requirements.txt
streamlit
opencv-python
numpy
Pillow
matplotlib
torch
torchvision
ultralytics
vietocr
